Fixed bug where Malware scan webhooks fired twice because we "pre-uploaded" a blob to lease. Simultaneously re-factored us out of the whole IBrokerStorage complex.
