+++ Test/Altinn.Broker.Tests/MalwareScanResultControllerTests.cs
@@ -9,8 +9,6 @@ using Altinn.Broker.Models;
 using Altinn.Broker.Tests.Factories;
 using Altinn.Broker.Tests.Helpers;
 
-using Microsoft.AspNetCore.TestHost;
-
 using Xunit;
 
 namespace Altinn.Broker.Tests;
@@ -57,7 +55,6 @@ public class MalwareScanResultControllerTests : IClassFixture<CustomWebApplicati
         // Get scanned status
         var scannedFile = await GetFileTransferWithNullAndOkCheck(fileTransferId);
         Assert.True(scannedFile.FileTransferStatus == FileTransferStatusExt.Failed);
-        Assert.True(scannedFile.FileTransferStatusText == "Malware scan failed");
     }
     [Fact]
     public async Task MalwareScanCreatedOneChangeOnDuplicateEvents()
@@ -91,7 +88,7 @@ public class MalwareScanResultControllerTests : IClassFixture<CustomWebApplicati
     private async Task<string> UploadAndCheckFileTransfer()
     {
         var initializeFileTransferResponse = await _senderClient.PostAsJsonAsync("broker/api/v1/filetransfer", FileTransferInitializeExtTestFactory.BasicFileTransfer());
-        Assert.Equal(System.Net.HttpStatusCode.OK, initializeFileTransferResponse.StatusCode);
+        Assert.True(initializeFileTransferResponse.IsSuccessStatusCode, $"The request failed with status code {initializeFileTransferResponse.StatusCode}. Error message: {await initializeFileTransferResponse.Content.ReadAsStringAsync()}");
         var fileTransferId = await initializeFileTransferResponse.Content.ReadAsStringAsync();
 
         // Upload
@@ -123,7 +120,7 @@ public class MalwareScanResultControllerTests : IClassFixture<CustomWebApplicati
     private async Task<HttpResponseMessage> SendMalvareScanResult(string jsonBody)
     {
         var result = await _webhookClient.PostAsync("broker/api/v1/webhooks/malwarescanresults", new StringContent(jsonBody, Encoding.UTF8, "application/json"));
-        Assert.Equal(System.Net.HttpStatusCode.OK, result.StatusCode);
+        Assert.True(result.IsSuccessStatusCode, $"The request failed with status code {result.StatusCode}. Error message: {await result.Content.ReadAsStringAsync()}");
         return result;
     }
 }
